refactor(act): extract validateRestoreRow to shared module
Three identical 25-line copies of the dry-run blocker validator in
each adapter (`in-memory-store.ts`, `postgres-store.ts`,
`sqlite-store.ts`) collapse to a single source in
`libs/act/src/restore-validate.ts`, exported publicly from
`@rotorsoft/act`. The validation rules ARE the dry-run contract —
keeping three copies in sync was a maintenance hazard for zero
gain.

Net change: -123 / +4 LOC across the four files, plus a new
80-line module with the function + canonical doc-comment.

External callers can now pre-validate a CSV before invoking
`restore` without recreating the validator — useful for CLI tools
that want a separate "lint backup" command.

## Stability charter impact

**Charter surface touched: `libs/act/src/index.ts`** — adds
`validateRestoreRow` to the public export list.

**Category: additive.** New public function, not a change to any
existing symbol. Per STABILITY.md's "additive port surface changes
go in `feat` commits and ship in minor releases" — but this is a
`refactor` commit rather than `feat` because the function existed
internally already; we're just promoting it from per-adapter
private to package-public. No `BREAKING CHANGE:` footer.
