+++ libs/act-sqlite/src/sqlite-store.ts
@@ -18,6 +18,7 @@ import type {
   StreamPosition,
   StreamStats,
 } from "@rotorsoft/act";
+import { validateRestoreRow } from "@rotorsoft/act";
 
 /**
  * SQLite store configuration
@@ -27,44 +28,6 @@ export interface SqliteConfig {
   url: string;
   /** Auth token for libSQL server connections (optional) */
   authToken?: string;
-}
-
-/**
- * Per-row blocker check for {@link SqliteStore.restore} dry-run mode.
- * Mirrored across `InMemoryStore` and `PostgresStore` — kept inlined
- * per adapter rather than exported so the public API stays narrow.
- * See `in-memory-store.ts:validateRestoreRow` for the canonical
- * comment.
- */
-function validateRestoreRow(
-  row: RestoreRow,
-  rowIdx: number,
-  seenIds: Set<number>,
-  expectedVersionByStream: Map<string, number>,
-  errors: Array<{ row: number; reason: string }>
-): void {
-  if (seenIds.has(row.id))
-    errors.push({ row: rowIdx, reason: `Duplicate id: ${row.id}` });
-  else seenIds.add(row.id);
-  if (row.version < 0)
-    errors.push({
-      row: rowIdx,
-      reason: `Negative version: ${row.version}`,
-    });
-  const created =
-    row.created instanceof Date ? row.created : new Date(row.created);
-  if (Number.isNaN(created.getTime()))
-    errors.push({
-      row: rowIdx,
-      reason: `Malformed created: ${String(row.created)}`,
-    });
-  const expected = expectedVersionByStream.get(row.stream) ?? 0;
-  if (row.version !== expected)
-    errors.push({
-      row: rowIdx,
-      reason: `Version gap on ${row.stream}: expected ${expected}, got ${row.version}`,
-    });
-  expectedVersionByStream.set(row.stream, row.version + 1);
 }
 
 const DEFAULT_CONFIG: SqliteConfig = {
