+++ libs/act/src/index.ts
@@ -10,5 +10,6 @@ export * from "./adapters/index.js";
 export * from "./builders/index.js";
 export * from "./config.js";
 export * from "./ports.js";
+export * from "./restore-validate.js";
 export * from "./types/index.js";
 export * from "./utils.js";
